1.01.004

+++ CAStoModbusRBT/Core/Inc/Modbus.h
@@ -38,7 +38,7 @@ typedef struct __attribute__((__packed__)) _modbusRxData {
 }modbusRxData_t;
 
 typedef struct __attribute__((__packed__))_modbusSettings {
-	uint8_t deviceId;
+	uint16_t deviceId;
 	uint16_t baudRate;
 }modbusSettings_t;
